+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Added [`Dropdown`](https://storybook.nulogy.design/?selectedKind=Dropdown&selectedStory=Dropdown) component
+- Added [`DropdownMenu`](https://storybook.nulogy.design/?selectedKind=DropdownMenu&selectedStory=DropdownMenu) component
 
 ### Changed
